privkey: check len and pointer in gnutls_x509_privkey_generate2
data_size could be greater than 0 and data could be NULL pointer in gnutls_x509_privkey_generate2 function.
data iteration is based on data_size, so date will be iterated if data_size is greater than 0 and data is
a valid pointer.
It iterates vdata if vdata_size is greater 0 and vdata is a valid pointer.
This issue was discovered using cppcheck tool.

Signed-off-by: Lucas Dias <lucas_dias@redes.ufsm.br>